Moved old bot functions into separate package
