+++ src/main/java/fi/metatavu/metamind/bot/functions/old/DateFromNowMetaBotFunction.java
@@ -1,4 +1,4 @@
-package fi.metatavu.metamind.bot.functions;
+package fi.metatavu.metamind.bot.functions.old;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
