+++ src/main/java/fi/metatavu/metamind/bot/functions/old/GetRandomSequencyItemMetaBotFunction.java
@@ -1,4 +1,4 @@
-package fi.metatavu.metamind.bot.functions;
+package fi.metatavu.metamind.bot.functions.old;
 
 import java.util.Map;
 import java.util.Random;
