+++ src/main/java/fi/metatavu/metamind/bot/functions/old/RoundMetaBotFunction.java
@@ -1,11 +1,10 @@
-package fi.metatavu.metamind.bot.functions;
+package fi.metatavu.metamind.bot.functions.old;
 
 import java.util.Map;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-import org.apache.commons.lang3.text.WordUtils;
 import org.slf4j.Logger;
 
 import com.bladecoder.ink.runtime.Story;
@@ -14,20 +13,20 @@ import com.rabidgremlin.mutters.core.IntentMatch;
 import com.rabidgremlin.mutters.core.session.Session;
 
 /**
- * Metabot function for capitalizing words in a String into capitalized words, that is each word is made up of a titlecase character and then a series of lowercase characters.
+ * Metabot function for rounding numbers.
  * 
  * @author Antti Leppä
  * @author Heikki Kurhinen
  */
 @ApplicationScoped
-public class CapitalizeFullyMetaBotFunction extends AbstractMetaBotFunction {
+public class RoundMetaBotFunction extends AbstractMetaBotFunction {
 
   @Inject
   private Logger logger;
 
   @Override
   public String getFunctionName() {
-    return "CAPITALIZE_FULLY";
+    return "ROUND";
   }
   
   @Override
@@ -35,26 +34,42 @@ public class CapitalizeFullyMetaBotFunction extends AbstractMetaBotFunction {
     Map<String, String> params = getParsedParam(paramString);
     
     String variableName = params.get("variable");
-    String delimiters = params.get("delimiters");
+    String nearValue = params.get("near");
     if (variableName == null) {
       logger.error("Could not capitalize without variable name");
       return;
     }
     
-    if (delimiters == null) {
-      delimiters = " ";
+    Integer near = null;
+    if (nearValue != null) {
+      try {
+        near = Integer.parseInt(nearValue);
+      } catch (NumberFormatException e) {
+        logger.warn("Failed to parse near value", e);
+      }
     }
     
-    String value = getVariableString(story, variableName);
+    Float value = getVariableNumber(story, variableName);
     if (value != null) {
-      String capitalized = WordUtils.capitalizeFully((String) value, delimiters.toCharArray());
+      Integer rounded = null;
+      
+      if (near != null) {
+        rounded = roundNear(value, near);
+      } else {
+        rounded = Math.round(value);
+      }
+      
       try {
-        setVariable(story, variableName, capitalized);
+        setVariable(story, variableName, rounded);
       } catch (Exception e) {
         logger.error("Could not set variable state", e);
         return;
       }
     }
+  }
+
+  private int roundNear(float value, int near) {
+    return Math.round((value / near)) * near;
   }
 
 }
